+++ pii-ner-exclude.txt
@@ -519,3 +519,4 @@ Parameter Store
 -1
 dontackemails2
 the Zendesk API's
+POST User
